stream/requirements.txt
binance-connector==3.5.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
websocket-client==1.7.0
websockets==12.0
loguru==0.7.2
